requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==3.0.0
constantly==23.10.4
cryptography==42.0.5
daphne==3.0.2
Django==4.2.7
django-apscheduler==0.6.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
mysql-connector-python==8.2.0
packaging==23.2
Pillow==10.1.0
protobuf==4.21.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
pytz==2024.1
PyYAML==6.0.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
whitenoise==6.6.0
zope.interface==6.2